type fix

+++ 0x0C-more_malloc_free/2-calloc.c
@@ -8,7 +8,7 @@
  *
  * Return: pointer to allocated memory, or NULL if fail
  */
-void *_calloc(unsigned int nmemb, size_t size)
+void *_calloc(unsigned int nmemb, unsigned int size)
 {
 	char *ptr;
 	size_t i;
